refactor: remove v666 suffixes
Rename pipeline modules and tests to drop the v666 suffix, updating imports and docs to match.

Clean up CLI/doc strings to remove v666 wording.
