+++ docs/architecture.md
@@ -1,17 +1,17 @@
 # Architecture
 
-Polylogue v666 is a small pipeline that ingests chat exports into SQLite, renders Markdown/HTML, and builds an FTS index.
+Polylogue is a small pipeline that ingests chat exports into SQLite, renders Markdown/HTML, and builds an FTS index.
 
 ## Core modules
 
 - `polylogue/config.py` and `polylogue/paths.py` define config schema, defaults, and XDG paths.
 - `polylogue/cli/click_app.py` is the CLI entrypoint and routes commands to the pipeline.
-- `polylogue/run_v666.py` coordinates planning, ingesting, rendering, and indexing.
+- `polylogue/run.py` coordinates planning, ingesting, rendering, and indexing.
 - `polylogue/source_ingest.py` loads local JSON/JSONL/ZIP exports and auto-detects provider formats.
 - `polylogue/drive_client.py` handles Google Drive OAuth + downloads; `polylogue/drive_ingest.py` turns Drive payloads into conversations.
 - `polylogue/store.py` and `polylogue/db.py` handle SQLite storage for conversations, messages, attachments, and runs.
-- `polylogue/render_v666.py` writes Markdown/HTML; `polylogue/assets.py` defines attachment paths.
-- `polylogue/index_v666.py` builds the SQLite FTS index; `polylogue/search_v666.py` performs queries.
+- `polylogue/render.py` writes Markdown/HTML; `polylogue/assets.py` defines attachment paths.
+- `polylogue/index.py` builds the SQLite FTS index; `polylogue/search.py` performs queries.
 - `polylogue/health.py` runs cached health checks.
 
 ## Data flow
