+++ polylogue/__init__.py
@@ -1,7 +1,7 @@
 __all__ = [
     "config",
     "db",
-    "run_v666",
-    "render_v666",
-    "search_v666",
+    "run",
+    "render",
+    "search",
 ]
