+++ tests/test_editor.py
